refactor[api-server]: LotteryService 에서 Parts 관련 로직을 PartsService 로 이관하도록 수정
