+++ src/main/java/com/watermelon/server/event/lottery/parts/service/PartsService.java
@@ -1,5 +1,6 @@
 package com.watermelon.server.event.lottery.parts.service;
 
+import com.watermelon.server.admin.dto.response.ResponseAdminPartsWinnerDto;
 import com.watermelon.server.event.lottery.parts.dto.response.ResponseMyPartsListDto;
 import com.watermelon.server.event.lottery.parts.dto.response.ResponsePartsDrawDto;
 import com.watermelon.server.event.lottery.parts.dto.response.ResponseRemainChanceDto;
@@ -32,5 +33,23 @@ public interface PartsService {
      * @return 링크에 대한 유저의 파츠 목록
      */
     List<ResponseMyPartsListDto> getPartsList(String linkKey);
+
+    /**
+     * 파츠 이벤트 응모 인원에 대해 뽑기를 진행한다.
+     */
+    void partsLottery();
+
+    /**
+     * 파츠 추첨 당첨자 명단을 반환
+     * @return 파츠 추첨 당첨자 리스트
+     */
+    List<ResponseAdminPartsWinnerDto> getAdminPartsWinners();
+
+
+    /**
+     * uid 에 해당하는 당첨자 파츠 추첨 어드민 확인 상태를 참으로 변경
+     * @param uid 당첨자의 uid
+     */
+    void partsWinnerCheck(String uid);
 
 }
