+++ src/main/java/com/watermelon/server/event/lottery/service/LotteryServiceImpl.java
@@ -33,7 +33,6 @@ public class LotteryServiceImpl implements LotteryService, LotteryWinnerService,
 
     private final LotteryApplierRepository lotteryApplierRepository;
     private final LotteryRewardRepository lotteryRewardRepository;
-    private final PartsRewardRepository partsRewardRepository;
 
     @Override
     public List<ResponseLotteryWinnerDto> getLotteryWinners() {
@@ -103,13 +102,6 @@ public class LotteryServiceImpl implements LotteryService, LotteryWinnerService,
                 .collect(Collectors.toList());
     }
 
-    @Override
-    public List<ResponseAdminPartsWinnerDto> getAdminPartsWinners() {
-        return lotteryApplierRepository.findByIsPartsWinnerTrue().stream()
-                .map(ResponseAdminPartsWinnerDto::from)
-                .collect(Collectors.toList());
-    }
-
     @Transactional //처음 상태와 변경 후 상태의 원자성 보장 필요.
     @Override
     public void lotteryWinnerCheck(String uid) {
@@ -118,13 +110,7 @@ public class LotteryServiceImpl implements LotteryService, LotteryWinnerService,
         lotteryApplierRepository.save(lotteryApplier);
     }
 
-    @Transactional //처음 상태와 변경 후 상태의 원자성 보장 필요.
-    @Override
-    public void partsWinnerCheck(String uid) {
-        LotteryApplier lotteryApplier = lotteryApplierRepository.findByUid(uid).orElseThrow();
-        lotteryApplier.partsWinnerCheck();
-        lotteryApplierRepository.save(lotteryApplier);
-    }
+
 
     @Transactional
     @Override
@@ -161,44 +147,6 @@ public class LotteryServiceImpl implements LotteryService, LotteryWinnerService,
 
     private List<LotteryApplier> getLotteryCandidates() {
         return lotteryApplierRepository.findByIsLotteryApplierTrue();
-    }
-
-    @Transactional
-    @Override
-    public void partsLottery() {
-        List<LotteryApplier> candidates = getPartsLotteryCandidates();
-        partsLotteryForCandidates(candidates);
-    }
-
-    private List<LotteryApplier> getPartsLotteryCandidates(){
-        return lotteryApplierRepository.findByIsPartsApplierTrue();
-    }
-
-    private void partsLotteryForCandidates(List<LotteryApplier> candidates) {
-
-        //참가자를 무작위로 섞는다.
-        Collections.shuffle(candidates);
-
-        //전체 보상 정보를 가져온다.
-        List<PartsReward> rewards = partsRewardRepository.findAll();
-
-        //당첨자를 저장할 리스트
-        List<LotteryApplier> lotteryWinners = new ArrayList<>();
-
-        int all_count=0;
-
-        //당첨 정보를 설정하고, 당첨자 인원만큼 리스트에 담는다.
-        for(PartsReward reward : rewards){
-            int winnerCount = reward.getWinnerCount();
-            for(int i=0; i<winnerCount; i++, all_count++){
-                LotteryApplier winner = candidates.get(all_count);
-                winner.partsLotteryWin();
-                lotteryWinners.add(winner);
-            }
-        }
-
-        lotteryApplierRepository.saveAll(lotteryWinners);
-
     }
 
     @Override
